refactor(#5): split account NDA sheet into hook, entry, and dialogs
Extract agreement entry, add/delete dialogs, and state hook so AccountDetailNdaPopover stays a thin sheet orchestrator.
