+++ app/dashboard/accounts/components/nda-delete-dialog.tsx
@@ -1,0 +1,60 @@
+'use client'
+
+import { Loader } from '@hugeicons/core-free-icons'
+
+import { AppIcon } from '@/lib/icons'
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+} from '@/components/ui/alert-dialog'
+
+export function NdaDeleteDialog({
+  open,
+  deleting,
+  onOpenChange,
+  onConfirm,
+}: {
+  open: boolean
+  deleting: boolean
+  onOpenChange: (open: boolean) => void
+  onConfirm: () => void
+}) {
+  return (
+    <AlertDialog
+      open={open}
+      onOpenChange={(v) => !v && !deleting && onOpenChange(false)}
+    >
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogTitle>NDA löschen?</AlertDialogTitle>
+          <AlertDialogDescription>
+            Die Vereinbarung und ein ggf. hochgeladenes PDF werden dauerhaft entfernt.
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+        <AlertDialogFooter>
+          <AlertDialogCancel disabled={deleting}>Abbrechen</AlertDialogCancel>
+          <AlertDialogAction
+            disabled={deleting}
+            className="bg-destructive text-destructive-foreground hover:bg-destructive/90"
+            onClick={(e) => {
+              e.preventDefault()
+              onConfirm()
+            }}
+          >
+            {deleting ? (
+              <AppIcon icon={Loader} size={16} className="animate-spin" />
+            ) : (
+              'Löschen'
+            )}
+          </AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
+  )
+}
